added `hashed` flag to the index on `motions.text`
This is to avoid 1024B limit on key size in MongoDB and the only way to
have an index on the `text` field of `motions` collection. The values are
occasionally longer there at least in sk/nrsr parliament. Range and regular
expression searches cannot be supported by a hashed index and therefore
will be slower. Though, at least some queries (the ones with equality test)
will be supported by the index and faster compared to have no index at
all.